reorganize and enhance windows console component to fail earlier allocating the screen
due to the selection between terminfo and console screens, we need to know at the time of selection whether each will work. actually since the terminfo screen tries to run first first, it's the one that this type of work (moving logic to alloc instead of init) is important for. however, I've done it for the windows console too, to move things in the right direction